requirements.txt
# Core numerical stack
numpy>=1.23
scipy>=1.10

# Deep learning
torch>=2.0

# Machine learning utilities
scikit-learn>=1.3

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Experiment tracking (optional)
tensorboard>=2.14

# Testing
pytest>=7.4

# Reproducibility
pyyaml>=6.0
